Complete main program
There is a bug that command 3 does not write all input file, will fix later

+++ bubble_sort.cpp
@@ -1,6 +1,6 @@
 #include "sorting_algorithm.h"
 
-void bubbleSort(vector<type>& v, int& comparision_count) {
+void bubbleSort(vector<type>& v, unsigned long long& comparision_count) {
 	int size = v.size();
 
 	for (int i = 0; ++comparision_count && i < size - 1; i++) {
